pyproject.toml
[tool.poetry]
name = "recruitai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.34.0"
pandas = "^2.2.2"
pymongo = "^4.7.2"
openpyxl = "^3.1.4"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
openai = "^1.40.6"
fuzzywuzzy = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
